- Restructure pipeline class
- Change uuid name fields to each respective id
